speech_to_text/requirements.txt
transformers==4.35.2
datasets==2.15.0
accelerate==0.25.0
click==8.1.7
tqdm==4.66.1
scipy==1.11.4
matplotlib==3.7.1
seaborn==0.12.2
numpy
pandas
joblib
openai-whisper==20231117
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.2.1